Split type-level `Ord` into a compatibility package.
This was added to `base` in 4.16.0.0, I don't see an existing
compatibility package, and I've already done the work to evolve the
prior implementation into a compatibility shim, so I may as well make it
usable without the rest of `numeric-kinds`.  This also breaks a
dependency cycle that recently appeared in other ongoing
`dependent-literals` work, which is the real motivation for doing this.
