+++ dependent-literals/src/DependentLiterals/Int.hs
@@ -92,7 +92,8 @@ import Data.SNumber (SafeSNumber, SNumber(N#), sameSNumber, unsafeMkSNumber)
 import Data.Tagged (Tagged(..))
 import Data.Wrapped (Wrapped(..))
 import Kinds.Integer (type (-#), KnownInteger(..), pattern Pos)
-import Kinds.Num (type (>=), type (<), ToInteger)
+import Kinds.Num (ToInteger)
+import Kinds.Ord (type (>=), type (<))
 import qualified Kinds.Integer as K (Integer)
 import Data.Fin.Int (Fin, finToInt, unsafeFin)
