+++ numeric-kinds/numeric-kinds.cabal
@@ -39,5 +39,6 @@ library
   hs-source-dirs:
     src
   build-depends:
-    base >=4.12 && <4.17
+    base >=4.12 && <4.17,
+    type-compare ==0.1.*
   default-language: Haskell2010
