+++ numeric-kinds/package.yaml
@@ -37,6 +37,7 @@ extra-source-files:
 
 dependencies:
   - base >= 4.12 && < 4.17
+  - type-compare >= 0.1 && < 0.2
 
 library:
   source-dirs: src
