+++ numeric-kinds/src/Kinds/Num.hs
@@ -31,21 +31,11 @@
 {-# LANGUAGE TypeOperators #-}
 
 module Kinds.Num
-         ( -- * Comparisons
-           Cmp, Compare
-         , type (<?), type (<=?), type (==?), type (/=?), type (>=?), type (>?)
-
-           -- ** Inequality Constraints
-         , type (<), type (<=), type (==), type (/=), type (>=), type (>)
-
-           -- * Conversions
-         , FromNat, ToInteger
+         ( -- * Conversions
+           FromNat, ToInteger
 
            -- * Arithmetic
          , type (+), type (-), type (*)
-
-           -- * Utility
-         , Proven, OrdCond
          ) where
 
 import Prelude hiding (Integer)
@@ -54,32 +44,6 @@ import GHC.TypeNats (Nat)
 import qualified GHC.TypeNats as N (type (+), type (-), type (*))
 
 import {-# source #-}  Kinds.Integer (Integer(..))
-
-#if MIN_VERSION_base(4, 16, 0)
-import Data.Type.Ord
-         ( Compare, OrdCond
-         , type (<?), type (>?), type (<=?), type (>=?)
-         , type (<=), type (>=), type (>)
-         )
-#else
-import GHC.TypeNats (CmpNat)
-
--- | Type-level Ord "kindclass".
---
--- Note this has an invisible dependent @k@ parameter that makes the
--- textually-identical instances for different kinds actually different.  Neat!
-type family Compare (x :: k) (y :: k) :: Ordering
-
--- That is, this is @Compare {k=Nat} x y@, which is distinct from
--- @Compare {k=Integer} x y@, even though both of them are written as
--- @Compare x y@.
-type instance Compare {- k=Nat -} x y = CmpNat x y
-#endif
-
--- | Backwards-compatibility alias for 'Compare'.
---
--- Prefer 'Compare' over this.
-type Cmp x y = Compare x y
 
 -- | Type-level numeric conversion from 'Nat'.  Like 'fromInteger' in 'Num'.
 type family FromNat (n :: Nat) :: k
@@ -105,47 +69,3 @@ type instance x - y = (N.-) x y
 type family (x :: k) * (y :: k) :: k
 
 type instance x * y = (N.*) x y
-
--- Recently added to base.
-#if !MIN_VERSION_base(4, 16, 0)
-
--- | Type-level eliminator for 'Ordering'.
---
--- @OrdCond o lt eq gt@ selects from among @lt@, @eq@, and @gt@ according
--- to @o@.
-type family OrdCond (o :: Ordering) (lt :: k) (eq :: k) (gt :: k) :: k where
-  OrdCond 'LT lt eq gt = lt
-  OrdCond 'EQ lt eq gt = eq
-  OrdCond 'GT lt eq gt = gt
-
-infix 4 <?, >?, <=?, >=?
-
-type x <?  y = OrdCond (Compare x y) True False False
-type x >?  y = OrdCond (Compare x y) False False True
-type x <=? y = OrdCond (Compare x y) True True False
-type x >=? y = OrdCond (Compare x y) False True True
-#endif
-
-infix ==?, /=?
-
-type x ==? y = OrdCond (Compare x y) False True False
-type x /=? y = OrdCond (Compare x y) True False True
-
--- | Turns a type-level 'Bool' into a 'Data.Kind.Constraint' that it's 'True'.
-type Proven b = b ~ 'True
-
--- Recently added to base.
-#if !MIN_VERSION_base(4, 16, 0)
-infix 4 <=, >=, >
-
-type x <= y = Proven (x <=? y)
-type x >= y = Proven (x >=? y)
-type x >  y = Proven (x >? y)
-#endif
-
-infix 4 <, ==, /=
-
--- The base-4.16.0.0 version of (<) is wrong.
-type x <  y = Proven (x <? y)
-type x == y = Proven (x ==? y)
-type x /= y = Proven (x /=? y)
