+++ snumber/package.yaml
@@ -38,6 +38,7 @@ extra-source-files:
 dependencies:
   - base >= 4.12 && < 4.17
   - numeric-kinds >= 0.1 && < 0.3
+  - type-compare >= 0.1 && < 0.2
 
 library:
   source-dirs: src
