+++ snumber/snumber.cabal
@@ -42,4 +42,5 @@ library
   build-depends:
       base >=4.12 && <4.17
     , numeric-kinds >=0.1 && <0.3
+    , type-compare ==0.1.*
   default-language: Haskell2010
