+++ type-compare/CHANGELOG.md
@@ -1,0 +1,8 @@
+# 0.1.0 (2021-11-04)
+
+Initial version, moved out of `numeric-kinds`.
+
+* `Compare` type family (re-exported from `base` when available).
+* Equality/inequality test operators.
+* Equality/inequality constraint operators.
+* `Proven` and `OrdCond` utility type families.
